requirements.txt
moviepy==1.0.3
openai==0.27.0
python-telegram-bot==13.0